+++ app/controllers/Controller.php
@@ -6,7 +6,9 @@ class Controller
 {
     public function view($viewPath, $data = []) : void
     {
+        //start session
         session_start();
+        
         //extract variables to be used in the view
         extract($data);
